backend: move the nvidia-drm and modeset checks earlier
Move the helper implementation and use further up/earlier the call
stack.
