+++ meson.build
@@ -48,7 +48,6 @@ endif
 
 sources = [
     'src/av1.c',
-    'src/backend-common.c',
     'src/export-buf.c',
     'src/direct/direct-export-buf.c',
     'src/direct/nv-driver.c',
